src/modules/users/services/UpdateProfileService.ts
import { injectable, inject } from 'tsyringe';
import AppErrro from '@shared/errors/AppError';
import IHashProvider from '@modules/users/providers/HashProvider/models/IHashProvider';
import IUsersRepository from '@modules/users/repositories/IUsersRepository';

import User from '@modules/users/infra/typeorm/entities/User';

interface IRequest {
    user_id: string;
    name: string;
    email: string;
    old_password?: string;
    password?: string;
}

@injectable()
class UpdateProfileService {
    constructor(
        @inject('UsersRepository')
        private usersRepository: IUsersRepository,

        @inject('HashProvider')
        private hashProvider: IHashProvider,
    ) {}

    public async execute({
        user_id,
        name,
        email,
        password,
        old_password,
    }: IRequest): Promise<User> {
        const user = await this.usersRepository.findById(user_id);

        if (!user) {
            throw new AppErrro('Usuário não encontrado.');
        }

        const userWithUpdatedEmail = await this.usersRepository.findByEmail(
            email,
        );

        if (userWithUpdatedEmail && userWithUpdatedEmail.id !== user_id) {
            throw new AppErrro('E-mail já em uso.');
        }

        user.name = name;
        user.email = email;

        if (password && !old_password) {
            throw new AppErrro(
                'Você precisa informar a senha antiga para definir uma nova senha',
            );
        }

        if (password && old_password) {
            const checkOldPassword = await this.hashProvider.compareHash(
                old_password,
                user.password,
            );

            if (!checkOldPassword) {
                throw new AppErrro('Senha antiga não confere.');
            }
        }

        if (password) {
            user.password = await this.hashProvider.generateHash(password);
        }

        return user;
    }
}

export default UpdateProfileService;
